Version update/bug fix
Added version informaation to the admin page, so that you can see the
current version at a glance. Also replaced some deprecated PHP code with
their updated replacements.
Fixed a bug where changing the default chatbot format in the admin page
did not change the value in the config file.