Script and DB Refactor/Multiple Bug Fixes
Added the last of the AIML tag functions/Refactored the DB/Multiple Bug
Fixes
1.) Added functions for the remaining AIML tags:
<thatstar>
<topicstar>
<gossip>
2.) Refactored the database, standardizing field names across all of the
tables.
3.) Refactored the admin pages, replacing $_GET, $_POST and $_REQUEST
with input
filtering functions.
4.) Fixed several minor bugs that were discovered during the database
refactoring
process.
5.) Re-designed the AIML tag functions, consolidating duplicate code
into a single
function. this cut more than 70 lines of code from the script while
retaining
exactly the same functionality and performance.
6.) Created an experimental config file editor to the admin folder, to
allow
botmasters to edit the file directly. For now, the script doesn't
actually
change the config file, but that functionality will be added in the
coming
weeks.
7.) Updated and modified the automatic upgrade script to incorporate all
of the
above listed changes.