Bug Fixes
Corrected a bug in the download script that caused the downloaded AIML
files to be saved with ASCII encoding, rather than the encoding set by
the config file.
Also cleaned up some more "orphaned" debugging code and made some minor
changes to the way the script's conversation array is initialized.