Conditional PDO support added, major bug fixes
1.) Added PDO support for PHP versions that support it (and have it
enabled),
with a fallback to the original MySQL functions if no PDO support is
detected
2.) Finally found the problem with template-side <that> tags not being
displayed
correctly. Special thanks to Tom (AKA Slow Putzo) for the assist with
this!
3.) Also fixed the long-standing bug with pattern-side <that> tags not
being
scored and chosen correctly. Once again, Thanks Tom! :D
4.) Removed the DB stats page from the admin. It was more or less a
failed
experiment, and really needed to be removed.
5.) Also removed the scripts to upgrade from version 2.0 to 2.3 - If you
have any
version of Program O that's older than 2.3.1, you need to do a fresh
install
and more or less start from scratch. Sorry, but that's just the way it
is.
6.) Other miscellaneous, minor bug fixes, mainly dealing with
unnecessary error
log entries.