Refactor for PDO support, refinement of srai_lookup
Cleaned up the remaining PDO calls that were causing errors, and changed
the default fetch mode to return associative arrays, thus reducing
memory requirements. Also altered the install script to look for the
srai_lookup table, and add it if necessary. This eliminates the need to
re-install to an empty DB. The script still does not add entries to the
srai lookup table when a usable candidate is found (that's next on the
agenda). I may write a separate script to insert them, and include it in
a "toolkit", or even add it to the admin pages. Still weighing options
there.