Bug fix, Feature change, minor style changes
Here we go, folks! A new toy to play with! But first, the stuff about
bugs:
1.) Fixed a minor bug where the correct chatbot wasn't being displayed
when changed until another admin page was loaded.
2.) Altered the behavior of the downloads page to allow for downloading
multiple files at once as a zip archive - It's not enough for a version
change, but I changed the build number. I'll be using the build numbers
more in the future.
3.) Updated the W3C validation icon, since Program O is now valid HTML5,
instead of XHTML 1.~whatever. There's now a sweet little "Valid HTML5"
icon in the lower left corner. YAY!
4.) Moved a function from the admin's index file to the common functions
file in the library. It just made more sense to have it there.