Relocated the session save path
Moved the session save path from the server default to a new folder.
this increases security for those on shared hosts, and ensures that
there are no permissions problems (or problems with the session save
path not being set in the PHP config) for those who have been
experiencing login problems due to sessions not being saved.