An Attempt to resolve issue #129
While investigating issue #129 , I noticed that there was no provision
for testing if an existing THATPATTERN failed to match an existing
previous bot response, so I added it in. Let's hope this helps. :)
I also fixed a small bug in the debug file viewer that was placing
entries in it's error log.