AIML Validation Change
1.) Changed the validation method in admin/validateAIML.php from using
XML DocType Declaration (DTD) o XML Schema (XSD). This allows the
validation script to check for improper characters in both `<pattern>`
and `<that>` tags, which has been a real issue in the past
2.) Incorporated the same validation method into the AIML upload script,
so that you are not now required to validate the files before uploading
them. You now cannot upload invalid AIML files.