Bug Fixes
Corrected some bugs that were causing fatal errors stemming from the
function session_gc() being commented out, but not the calls to it. Also
streamlined the code a little in the file find_aiml.php and added some
experimental code to provide some improved formatting for arrays and
objects in the debug files.