New Feature, Improved Search, and Bug fixes
Added something that was in the works since version 2.0, but never
implemented; Unknown inputs! Now, whenever either the default AIML
cagegory (usually RANDOM PICKUP LINE) is called, or if the default error
response is used, the input that triggered it is stored in a special
table so that it can be reviewed later. An admin page was also added to
work with these unknown inputs, giving botmasters the chance to improve
their AIML code much more easily.

Also, a new search method has been implemented that both streamlines the
main search SQL query, making it easier to read, it also catches more
potential matches, reducing the number of "misses" considerably.

In addition to the improved search query, some bugs were discovered that
were causing topics to be missed. This will also improve search
accuracy.