Added function db_parseSQL
Added a new function to the DB handlers to parse prepared statement type
SQL queries for outputting to logs. The function takes the SQL query and
the array of replacement parameters and converts the query to something
that can be read by people, and can be directly copied/pasted into DB
management software, such as phpMyAdmin for testing purposes.

Please note that at this time the function is not actually being used.
That will come later.