Fixed issue #280
I noticed with the help of @rwaters71 that the learn tag
was not completly implemented. Because the function
`find_user_defined aiml` and `find_aiml_matches` are mostly the same,
I merged them together.

Plus fixed a bug where some the fields `pattern`, `thatpattern` and
`topic` were not displayed correctly in the
AIML editor. They will now also show the bot tag.